Sentry/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.0
joblib==1.4.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
matplotlib==3.9.0
instaloader>=4.10
sqlalchemy>=2.0.0
